# specs/004-security-compliance/contracts/security-admin-openapi.yaml
openapi: 3.1.0
info:
  title: PowerX Plugin Security & Compliance API
  version: 0.1.0
  description: >
    Admin and agent endpoints that implement privacy, ToolGrant, and vulnerability
    response requirements for the PowerX base plugin.
servers:
  - url: https://{host}/_p/{pluginId}/api/v1
    variables:
      host:
        default: tenant.powerx.host
      pluginId:
        default: com.powerx.plugin.base
paths:
...
  /admin/security/advisories:
    post:
      summary: Create vulnerability advisory
      operationId: createAdvisory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdvisoryBody"
      responses:
        "201":
          description: Advisory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnerabilityAdvisory"
    get:
      summary: List advisories
      operationId: listAdvisories
      security:
        - bearerAuth: []
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW]
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, PATCHED, PUBLISHED, CLOSED]
      responses:
        "200":
          description: Advisories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnerabilityAdvisoryList"
  /admin/security/advisories/{advisoryId}/publish:
    post:
      summary: Publish advisory and trigger notifications
      operationId: publishAdvisory
      security:
        - bearerAuth: []
      parameters:
        - name: advisoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishAdvisoryBody"
      responses:
        "200":
          description: Advisory published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnerabilityAdvisory"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreateAdvisoryBody:
      type: object
      required: [reference, severity, summary, affected_versions]
      properties:
        reference:
          type: string
        severity:
          type: string
        summary:
          type: string
        details_markdown:
          type: string
        affected_versions:
          type: array
          items:
            type: string
        sla_deadline:
          type: string
          format: date-time
    PublishAdvisoryBody:
      type: object
      required: [patched_in_version]
      properties:
        patched_in_version:
          type: string
        notify_channels:
          type: array
          items:
            type: string
    VulnerabilityAdvisory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        severity:
          type: string
        status:
          type: string
        affected_versions:
          type: array
          items:
            type: string
        patched_in_version:
          type: string
          nullable: true
        summary:
          type: string
        published_at:
          type: string
          format: date-time
          nullable: true
    VulnerabilityAdvisoryList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/VulnerabilityAdvisory"
...